pyproject.toml
[tool.poetry]
name = "commonroad-dataset-converter"
version = "2023.2"
description = "CommonRoad Dataset Converter"
authors = ["Cyber-Physical Systems Group, Technical University of Munich <commonroad@lists.lrz.de>"]
license = "BSD-3-Clause"
readme = "README.md"
packages = [{ include = "commonroad_dataset_converter" }]
homepage = "https://commonroad.in.tum.de"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
exclude = ["**/*.ipynb"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21"
scipy = ">=1.4.1"
pandas = ">=0.24.2"
ruamel-yaml = ">=0.16.10"
commonroad-io = ">=2023.2"
typer = ">=0.4.0"
tqdm = "^4.64.1"
pyproj = "^3.4.0"
pyarrow = "^9.0.0"
commonroad-route-planner = "^2022.3"
commonroad-drivability-checker = "^2023.1"
pydantic = "^2.3.0"

[tool.poetry.scripts]
crconvert = 'commonroad_dataset_converter.main:cli'

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[tool.poetry.group.doc]
optional = true

[tool.poetry.group.doc.dependencies]
sphinx = "^5.3.0"
myst-parser = "^1.0.0"
sphinx-book-theme = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
